# Optimize underscore inflector